refactor(kolme): extract http transport module ownership (#1954)
